requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.28.25
botocore==1.31.25
Django==4.2.4
django-filter==23.2
django-storages==1.13.2
graphene==3.3
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
jmespath==1.0.1
promise==2.3
psycopg2==2.9.7
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==1.26.16